Fixed: Issue with game/ai - follow player
Because they only checked distance to $player not to the actual follow target the ai did correctly follow the player.
But with the previous upgrade we check now the actual distance to the actual follow target.

The FOLLOW_TARGET_OUT_OF_RANGE check requires that followtarget is explicitly set, which they omitted in the scripts.

So we need to set the followtarget reliably some how, and the best way we can do this in a controlled manner is via the state/ai file.