Disabled Local/Listen Server in Menu for Starting a Coop Server
Because of on going memory issues that result in visual glitches in the client/engine code listen server option has been disabled.

We can not fix that at this time since we don't have the code.
We can also not pinpoint it exactly and workarounds don't work right on listen servers.

Flushtikis doesn't do the Job to the extend we would require it to.
Also, textures are retained on the client when hosting, eventually leading to a max images (something) error.